\docType{data}
\name{bowelcancer}
\alias{bowelcancer}
\title{Bowel cancer mortality}
\format{A data frame with 379 rows and 3 variables}
\source{
  Scraped from

  Goldacre, Ben (28th October 2011) "DIY statistical
  analysis: experience the thrill of touching real data",
  The Guardian,
  \url{http://www.guardian.co.uk/commentisfree/2011/oct/28/bad-science-diy-data-analysis}.
}
\description{
  Bowel cancer mortality rates by UK local authority.
}
\details{
  The variables are as follows:

  \itemize{ \item population.  Total population \item
  deaths.  Bowel cancer deaths \item rate.  Deaths per
  100,000 }
}
\keyword{datasets}

