% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_specs.R
\docType{data}
\name{default_text_specs}
\alias{default_text_specs}
\title{Default specifications for the \code{ggplot2::geom_text()} geom}
\format{
An object of class \code{list} of length 16.
}
\usage{
default_text_specs
}
\description{
\code{ggfunnel::default_text_specs} store the list of default values
used at the \code{ggplot2::geom_text()} geom arguments.
}
\details{
If the argument \code{labels} of \code{ggfunnel::funnel()} is set to \code{TRUE}, than,
the \code{ggplot2::geom_text()} geom will be used to print the value labels in the funnel chart.
\code{ggfunnel::funnel()} will fill all arguments of \code{ggplot2::geom_text()} geom with the values
present at \code{ggfunnel::default_text_specs}.

You can change these values by using the \code{text_specs} argument of \code{ggfunnel::funnel()}.
}
\keyword{datasets}
