% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_specs.R
\docType{data}
\name{default_tile_specs}
\alias{default_tile_specs}
\title{Default specifications for the \code{ggplot2::geom_tile()} geom}
\format{
An object of class \code{list} of length 11.
}
\usage{
default_tile_specs
}
\description{
\code{ggfunnel::default_tile_specs} store the list of default values
used at the \code{ggplot2::geom_tile()} geom arguments.
}
\details{
\code{ggfunnel::funnel()} uses the
\code{ggplot2::geom_tile()} geom to build the rectangles present in the funnel chart,
and will fill all arguments of \code{ggplot2::geom_tile()} geom with the values
present at \code{ggfunnel::default_tile_specs}.

You can change these values by using the \code{tile_specs} argument of \code{ggfunnel::funnel()}.
}
\keyword{datasets}
