% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel.R
\name{funnel}
\alias{funnel}
\title{Plot a Funnel chart}
\usage{
funnel(data, values, levels)
}
\arguments{
\item{data}{The data.frame with your data;}

\item{values}{The column name where are the values you want to display in your funnel chart;}

\item{levels}{The column name with the "levels" (or the "groups") you want to
display in your funnel chart;}
}
\description{
Plot a Funnel chart
}
